Limit max idle connections to 1 for http client
This does not seem to affect results.

After diggin through tonic, tower and hyper source code, it seems that
the tonic default gRPC client uses a single http2 connection to the
server. I suspected the reqwest client in the http/rest variant to
utilize multiple http2 connections. However, it seems to only do so when
the reqwest client is used against multiple host.

The change did not affect http/rest results negatively.